HBT-38 Move methods related to data aggregation to new service
