+++ src/test/java/com/rainy/homebudgettracker/transaction/TransactionServiceTest.java
@@ -13,6 +13,7 @@ import com.rainy.homebudgettracker.images.ImageService;
 import com.rainy.homebudgettracker.mapper.ModelMapper;
 import com.rainy.homebudgettracker.transaction.enums.CurrencyCode;
 import com.rainy.homebudgettracker.transaction.enums.TransactionMethod;
+import com.rainy.homebudgettracker.transaction.service.TransactionServiceImpl;
 import com.rainy.homebudgettracker.user.UserService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -102,16 +103,6 @@ class TransactionServiceTest {
         when(modelMapper.map(TestData.CONVERTED_TRANSACTION_3, TransactionResponse.class)).thenReturn(TestData.CONVERTED_TRANSACTION_RESPONSE_3);
         when(transactionRepository.save(TestData.CONVERTED_TRANSACTION_3)).thenReturn(TestData.CONVERTED_TRANSACTION_3);
         when(exchangeService.getExchangeRate(CurrencyCode.EUR, CurrencyCode.PLN)).thenReturn(ExchangeResponse.builder().conversionRate("4.22").build());
-
-        when(transactionRepository.sumPositiveAmountByAccount(TestData.ACCOUNT)).thenReturn(BigDecimal.valueOf(100.1));
-        when(transactionRepository.sumPositiveAmountByAccount(TestData.ACCOUNT_2)).thenReturn(BigDecimal.valueOf(0));
-        when(transactionRepository.sumNegativeAmountByAccount(TestData.ACCOUNT)).thenReturn(BigDecimal.valueOf(100.1));
-        when(transactionRepository.sumNegativeAmountByAccount(TestData.ACCOUNT_2)).thenReturn(BigDecimal.valueOf(0));
-        when(transactionRepository.sumAmountByAccount(TestData.ACCOUNT)).thenReturn(BigDecimal.valueOf(100.1));
-        when(transactionRepository.sumAmountByAccount(TestData.ACCOUNT_2)).thenReturn(BigDecimal.valueOf(0));
-
-        when(modelMapper.map(BigDecimal.valueOf(100.1).setScale(2, RoundingMode.HALF_UP), SumResponse.class)).thenReturn(TestData.SUM_RESPONSE);
-        when(modelMapper.map(BigDecimal.valueOf(0).setScale(2, RoundingMode.HALF_UP), SumResponse.class)).thenReturn(TestData.SUM_RESPONSE_2);
 
         when(transactionRepository.findAllByUserSub(TestData.USER_SUB)).thenReturn(List.of(TestData.TRANSACTION));
     }
@@ -330,49 +321,6 @@ class TransactionServiceTest {
     void shouldThrowExceptionWhenUserIsNotOwnerWhenDeleting() {
         assertThrows(UserIsNotOwnerException.class,
                 () -> transactionService.deleteCurrentUserTransaction(UUID.fromString(TestData.TRANSACTION_ID_2)));
-    }
-
-    @Test
-    void shouldReturnSumPositiveAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumPositiveAmount = transactionService.sumCurrentUserPositiveAmount(UUID.fromString(TestData.ACCOUNT_ID));
-
-        assertEquals(TestData.SUM_RESPONSE, sumPositiveAmount);
-    }
-
-    @Test
-    void shouldReturnSumAs0PositiveAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumPositiveAmount = transactionService.sumCurrentUserPositiveAmount(UUID.fromString(TestData.ACCOUNT_ID_2));
-
-        assertEquals(TestData.SUM_RESPONSE_2, sumPositiveAmount);
-
-    }
-
-    @Test
-    void shouldReturnSumNegativeAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumNegativeAmount = transactionService.sumCurrentUserNegativeAmount(UUID.fromString(TestData.ACCOUNT_ID));
-
-        assertEquals(TestData.SUM_RESPONSE, sumNegativeAmount);
-    }
-
-    @Test
-    void shouldReturnSumAs0NegativeAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumPositiveAmount = transactionService.sumCurrentUserNegativeAmount(UUID.fromString(TestData.ACCOUNT_ID_2));
-
-        assertEquals(TestData.SUM_RESPONSE_2, sumPositiveAmount);
-    }
-
-    @Test
-    void shouldReturnSumAllAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumAllAmount = transactionService.sumCurrentUserAmount(UUID.fromString(TestData.ACCOUNT_ID));
-
-        assertEquals(TestData.SUM_RESPONSE, sumAllAmount);
-    }
-
-    @Test
-    void shouldReturnSumAs0AllAmount() throws RecordDoesNotExistException, UserIsNotOwnerException {
-        var sumPositiveAmount = transactionService.sumCurrentUserAmount(UUID.fromString(TestData.ACCOUNT_ID_2));
-
-        assertEquals(TestData.SUM_RESPONSE_2, sumPositiveAmount);
     }
 
     @Test
